# api/api.yaml
openapi: 3.0.2
info:
  version: v1
  title: BAI2 API
  description: Moov Bai2 ([Automated Clearing House](https://en.wikipedia.org/wiki/Automated_Clearing_House)) implements an HTTP API for creating, parsing and validating Bais files. BAI2- a widely accepted and used Bank Statement Format for Bank Reconciliation.
  contact:
    url: https://github.com/moov-io/bai2
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

servers:
  - url: http://localhost:8208
    description: Local development

tags:
  - name: 'Bai2 Files'
    description: |
      BAI2 file format is a standardized set of codes which comes in text format.

paths:
  /health:
    get:
      tags: ['Bai2 Files']
      summary: health bai2 service
      description: Check the bai2 service to check if running
      operationId: health
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: 'health: true'

  /print:
    post:
      tags: ['Bai2 Files']
      summary: Print bai2 file after parse bin file
      description: Print bai2 file.
      operationId: print
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                input:
                  type: string
                  description: bai2 bin file
                  format: binary
            encoding:
              file:
                contentType: text/plain
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: '01,0004,12345,060321,0829,001,80,1,2/
                
02,12345,0004,1,060317,,CAD,/

03,10200123456,CAD,040,+000000000000,,,045,+000000000000,,/

88,100,000000000208500,00003,V,060316,,400,000000000208500,00008,V,060316,/

16,409,000000000002500,V,060316,,,,RETURNED CHEQUE     /

16,409,000000000090000,V,060316,,,,RTN-UNKNOWN         /

16,409,000000000000500,V,060316,,,,RTD CHQ SERVICE CHRG/

16,108,000000000203500,V,060316,,,,TFR 1020 0345678    /

16,108,000000000002500,V,060316,,,,MACLEOD MALL        /

16,108,000000000002500,V,060316,,,,MASCOUCHE QUE       /

16,409,000000000020000,V,060316,,,,1000 ISLANDS MALL   /

16,409,000000000090000,V,060316,,,,PENHORA MALL        /

16,409,000000000002000,V,060316,,,,CAPILANO MALL       /

16,409,000000000002500,V,060316,,,,GALERIES LA CAPITALE/

16,409,000000000001000,V,060316,,,,PLAZA ROCK FOREST   /

49,+00000000000834000,000000014/

03,10200123456,CAD,040,+000000000000,,,045,+000000000000,,/

88,100,000000000111500,00002,V,060317,,400,000000000111500,00004,V,060317,/

16,108,000000000011500,V,060317,,,,TFR 1020 0345678    /

16,108,000000000100000,V,060317,,,,MONTREAL            /

16,409,000000000100000,V,060317,,,,GRANDFALL NB        /

16,409,000000000009000,V,060317,,,,HAMILTON ON         /

16,409,000000000002000,V,060317,,,,WOODSTOCK NB        /

16,409,000000000000500,V,060317,,,,GALERIES RICHELIEU  /

49,+00000000000446000,000000009/

98,+00000000001280000,000000002,000000025/

99,+00000000001280000,000000001,000000027/'
        '400':
          description: request
          content:
            text/plain:
              schema:
                type: string
                example: invalid file format

  /parse:
    post:
      tags: ['Bai2 Files']
      summary: Parse bai2 file after parse bin file
      description: parse bai2 file.
      operationId: parse
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                input:
                  type: string
                  description: bai2 bin file
                  format: binary
            encoding:
              file:
                contentType: text/plain
      responses:
        '200':
          description: successful operation
          content:
            text/plain:
              schema:
                type: string
                example: valid file
        '400':
          description: request
          content:
            text/plain:
              schema:
                type: string
                example: invalid file format

  /format:
    post:
      tags: ['Bai2 Files']
      summary: Format bai2 file after parse bin file
      description: format bai2 file.
      operationId: format
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                input:
                  type: string
                  description: bai2 bin file
                  format: binary
            encoding:
              file:
                contentType: text/plain
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: object
        '400':
          description: request
          content:
            text/plain:
              schema:
                type: string
                example: invalid file format